+++ .gitignore
@@ -15,3 +15,5 @@
 docs/doctrees
 docs/api
 docs/_build
+
+HSTB/tests/resources
