+++ HSTB/tests/modules/module_test_arrays.py
@@ -1,22 +1,4 @@
-import pytest
-
-from HSTB.kluster.modules.orientation import *
-from HSTB.kluster.modules.beampointingvector import *
-from HSTB.kluster.modules.svcorrect import *
-from HSTB.kluster.modules.georeference import *
-from HSTB.kluster.modules.georeference import _new_geohash
-from HSTB.kluster.modules.tpu import *
-
-try:
-    from hstb_kluster.tests.test_fqpr_generation import load_dataset
-except ImportError:
-    # relative import as tests directory can vary in location depending on how kluster is installed
-    from .test_fqpr_generation import load_dataset
-
-try:
-    from hstb_kluster.tests.test_datasets import RealFqpr
-except ImportError:
-    from .test_datasets import RealFqpr
+import numpy as np
 
 # tx vector is a 3d vector representing the transmitter orientation for each beam
 tx_vector = np.array([0.6136555921172974, -0.7895255928982701, 0.008726535498373935])
@@ -1058,280 +1040,3 @@ expected_thu = np.array([[4.7841897, 4.4138813, 4.6790714, 4.2010245, 4.733156, 
                           4.7419972, 4.773464, 4.191487, 4.1644616]], dtype=np.float32)
 
 
-def test_orientation_module():
-    dset = load_dataset(RealFqpr())
-    raw_attitude = dset.raw_att
-    # expand_dims required to maintain the time dimension metadata when you select only one value
-    multibeam = dset.raw_ping[0].isel(time=0).expand_dims('time')
-    traveltime = multibeam.traveltime
-    delay = multibeam.delay
-    timestamps = multibeam.time
-
-    installation_params_time = list(dset.xyzrph['tx_r'].keys())[0]
-    tx_orientation = [np.array([1, 0, 0]),  # starting vector for the tx transducer (points forward)
-                      dset.xyzrph['tx_r'][installation_params_time],  # roll mounting angle for tx
-                      dset.xyzrph['tx_p'][installation_params_time],  # pitch mounting angle for tx
-                      dset.xyzrph['tx_h'][installation_params_time],  # yaw mounting angle for tx
-                      installation_params_time]  # time stamp for the installation parameters record
-    rx_orientation = [np.array([0, 1, 0]),  # same but for the receiver
-                      dset.xyzrph['rx_r'][installation_params_time],
-                      dset.xyzrph['rx_p'][installation_params_time],
-                      dset.xyzrph['rx_h'][installation_params_time],
-                      installation_params_time]
-    latency = 0  # no latency applied for this test
-
-    calc_tx_vector, calc_rx_vector = build_orientation_vectors(raw_attitude, traveltime, delay, timestamps,
-                                                               tx_orientation, rx_orientation, latency)
-
-    try:
-        assert np.array_equal(calc_tx_vector.values, expected_tx_vector)
-    except AssertionError:
-        print('Falling back to approx, should only be seen in TravisCI environment in my experience')
-        # use approx here, I get ever so slightly different answers in the Travis CI environment
-        assert calc_tx_vector.values == pytest.approx(expected_tx_vector, 0.000001)
-    try:
-        assert np.array_equal(calc_rx_vector.values, expected_rx_vector)
-    except AssertionError:
-        print('Falling back to approx, should only be seen in TravisCI environment in my experience')
-        # use approx here, I get ever so slightly different answers in the Travis CI environment
-        assert calc_rx_vector.values == pytest.approx(expected_rx_vector, 0.000001)
-
-
-def test_beampointingvector_module():
-    dset = load_dataset(RealFqpr())
-    raw_attitude = dset.raw_att
-    heading = raw_attitude.heading
-
-    multibeam = dset.raw_ping[0].isel(time=0).expand_dims('time')
-    beampointingangle = multibeam.beampointingangle
-    tiltangle = multibeam.tiltangle
-    ping_time_heading = heading.interp_like(beampointingangle)
-
-    tx_vecs = xr.DataArray(data=expected_tx_vector, dims=['time', 'beam', 'xyz'], coords={'time': multibeam.time.values,
-                                                                                          'beam': multibeam.beam.values,
-                                                                                          'xyz': ['x', 'y', 'z']})
-    rx_vecs = xr.DataArray(data=expected_rx_vector, dims=['time', 'beam', 'xyz'], coords={'time': multibeam.time.values,
-                                                                                          'beam': multibeam.beam.values,
-                                                                                          'xyz': ['x', 'y', 'z']})
-
-    tx_reversed = False
-    rx_reversed = False
-
-    beam_azimuth, corrected_beam_angle = build_beam_pointing_vectors(ping_time_heading, beampointingangle, tiltangle,
-                                                                     tx_vecs,
-                                                                     rx_vecs, tx_reversed, rx_reversed)
-
-    try:
-        assert np.array_equal(beam_azimuth.values, expected_beam_azimuth)
-    except AssertionError:
-        print('Falling back to approx, should only be seen in TravisCI environment in my experience')
-        # use approx here, I get ever so slightly different answers in the Travis CI environment
-        assert beam_azimuth.values == pytest.approx(expected_beam_azimuth, 0.000001)
-    try:
-        assert np.array_equal(corrected_beam_angle.values, expected_corrected_beam_angles)
-    except AssertionError:
-        print('Falling back to approx, should only be seen in TravisCI environment in my experience')
-        # use approx here, I get ever so slightly different answers in the Travis CI environment
-        assert corrected_beam_angle.values == pytest.approx(expected_corrected_beam_angles, 0.000001)
-
-
-def test_svcorrect_module():
-    dset = load_dataset(RealFqpr())
-    multibeam = dset.raw_ping[0].isel(time=0).expand_dims('time')
-    cast = json.loads(multibeam.profile_1495599960)
-    beam_azimuth = xr.DataArray(data=expected_beam_azimuth, dims=['time', 'beam'],
-                                coords={'time': multibeam.time.values,
-                                        'beam': multibeam.beam.values})
-    beam_angle = xr.DataArray(data=expected_corrected_beam_angles, dims=['time', 'beam'],
-                              coords={'time': multibeam.time.values,
-                                      'beam': multibeam.beam.values})
-    traveltime = multibeam.traveltime
-    surface_ss = multibeam.soundspeed
-
-    installation_params_time = list(dset.xyzrph['tx_r'].keys())[0]
-    waterline = float(dset.xyzrph['waterline'][installation_params_time])
-    additional_offsets = np.array([[0], [0], [0]])
-
-    alongtrack, acrosstrack, depth = run_ray_trace_v2(cast, beam_azimuth, beam_angle, traveltime, surface_ss, waterline,
-                                                      additional_offsets)
-    assert np.array_equal(alongtrack, expected_alongtrack)
-    assert np.array_equal(acrosstrack, expected_acrosstrack)
-    assert np.array_equal(depth, expected_depth)
-
-
-def test_georeference_module():
-    dset = load_dataset(RealFqpr())
-    multibeam = dset.raw_ping[0].isel(time=0).expand_dims('time')
-    x = xr.DataArray(data=expected_alongtrack, dims=['time', 'beam'], coords={'time': multibeam.time.values,
-                                                                              'beam': multibeam.beam.values})
-    y = xr.DataArray(data=expected_acrosstrack, dims=['time', 'beam'], coords={'time': multibeam.time.values,
-                                                                               'beam': multibeam.beam.values})
-    z = xr.DataArray(data=expected_depth, dims=['time', 'beam'], coords={'time': multibeam.time.values,
-                                                                         'beam': multibeam.beam.values})
-    sv_corr = [x, y, z]
-
-    raw_attitude = dset.raw_att
-    heading = raw_attitude.heading.interp_like(z)
-    heave = raw_attitude.heave.interp_like(z)
-
-    altitude = dset.raw_ping[0].altitude
-    longitude = dset.raw_ping[0].longitude
-    latitude = dset.raw_ping[0].latitude
-
-    installation_params_time = list(dset.xyzrph['tx_r'].keys())[0]
-    waterline = float(dset.xyzrph['waterline'][installation_params_time])
-    vert_ref = 'waterline'
-
-    input_datum = CRS.from_epsg(7911)
-    output_datum = CRS.from_epsg(26910)
-
-    z_offset = 1.0
-
-    georef_x, georef_y, georef_z, corrected_heave, corrected_altitude, vdatumunc, geohashes = georef_by_worker(sv_corr, altitude,
-                                                                                                               longitude, latitude,
-                                                                                                               heading, heave,
-                                                                                                               waterline, vert_ref,
-                                                                                                               input_datum,
-                                                                                                               output_datum,
-                                                                                                               z_offset)
-    assert np.array_equal(georef_x, expected_georef_x)
-    assert np.array_equal(georef_y, expected_georef_y)
-    assert np.array_equal(georef_z, expected_georef_z)
-
-
-def test_georef_with_nan():
-    dset = load_dataset(RealFqpr())
-    multibeam = dset.raw_ping[0].isel(time=0).expand_dims('time')
-    at_with_nan = expected_alongtrack.copy()
-    at_with_nan[0][10] = np.nan
-    at_with_nan[0][20] = np.nan
-    ac_with_nan = expected_acrosstrack.copy()
-    ac_with_nan[0][10] = np.nan
-    ac_with_nan[0][20] = np.nan
-
-    x = xr.DataArray(data=at_with_nan, dims=['time', 'beam'], coords={'time': multibeam.time.values,
-                                                                      'beam': multibeam.beam.values})
-    y = xr.DataArray(data=ac_with_nan, dims=['time', 'beam'], coords={'time': multibeam.time.values,
-                                                                      'beam': multibeam.beam.values})
-    z = xr.DataArray(data=expected_depth, dims=['time', 'beam'], coords={'time': multibeam.time.values,
-                                                                         'beam': multibeam.beam.values})
-    sv_corr = [x, y, z]
-
-    raw_attitude = dset.raw_att
-    heading = raw_attitude.heading.interp_like(z)
-    heave = raw_attitude.heave.interp_like(z)
-
-    altitude = dset.raw_ping[0].altitude
-    longitude = dset.raw_ping[0].longitude
-    latitude = dset.raw_ping[0].altitude
-
-    installation_params_time = list(dset.xyzrph['tx_r'].keys())[0]
-    waterline = float(dset.xyzrph['waterline'][installation_params_time])
-    vert_ref = 'waterline'
-
-    input_datum = CRS.from_epsg(7911)
-    output_datum = CRS.from_epsg(26910)
-
-    z_offset = 1.0
-
-    georef_x, georef_y, georef_z, corrected_heave, corrected_altitude, vdatumunc, geohashes = georef_by_worker(sv_corr, altitude,
-                                                                                                               longitude, latitude,
-                                                                                                               heading, heave,
-                                                                                                               waterline, vert_ref,
-                                                                                                               input_datum,
-                                                                                                               output_datum,
-                                                                                                               z_offset)
-    assert np.isnan(georef_x.values[0][10])
-    assert np.isnan(georef_x.values[0][20])
-    assert np.isnan(georef_y.values[0][10])
-    assert np.isnan(georef_y.values[0][20])
-    assert np.array_equal(georef_z, expected_georef_z)
-
-
-def test_georef_with_depth_nan():
-    dset = load_dataset(RealFqpr())
-    multibeam = dset.raw_ping[0].isel(time=0).expand_dims('time')
-    dpth_with_nan = expected_depth.copy()
-    dpth_with_nan[0][10] = np.nan
-    dpth_with_nan[0][20] = np.nan
-
-    x = xr.DataArray(data=expected_alongtrack, dims=['time', 'beam'], coords={'time': multibeam.time.values,
-                                                                              'beam': multibeam.beam.values})
-    y = xr.DataArray(data=expected_acrosstrack, dims=['time', 'beam'], coords={'time': multibeam.time.values,
-                                                                               'beam': multibeam.beam.values})
-    z = xr.DataArray(data=dpth_with_nan, dims=['time', 'beam'], coords={'time': multibeam.time.values,
-                                                                        'beam': multibeam.beam.values})
-    sv_corr = [x, y, z]
-
-    raw_attitude = dset.raw_att
-    heading = raw_attitude.heading.interp_like(z)
-    heave = raw_attitude.heave.interp_like(z)
-
-    altitude = dset.raw_ping[0].altitude
-    longitude = dset.raw_ping[0].longitude
-    latitude = dset.raw_ping[0].latitude
-
-    installation_params_time = list(dset.xyzrph['tx_r'].keys())[0]
-    waterline = float(dset.xyzrph['waterline'][installation_params_time])
-    vert_ref = 'waterline'
-
-    input_datum = CRS.from_epsg(7911)
-    output_datum = CRS.from_epsg(26910)
-
-    z_offset = 1.0
-
-    georef_x, georef_y, georef_z, corrected_heave, corrected_altitude, vdatumunc, geohashes = georef_by_worker(sv_corr, altitude,
-                                                                                                               longitude, latitude,
-                                                                                                               heading, heave,
-                                                                                                               waterline, vert_ref,
-                                                                                                               input_datum,
-                                                                                                               output_datum,
-                                                                                                               z_offset)
-    assert np.array_equal(georef_x, expected_georef_x)
-    assert np.array_equal(georef_y, expected_georef_y)
-    assert np.isnan(georef_z.values[0][10])
-    assert np.isnan(georef_z.values[0][20])
-
-
-def test_tpu():
-    dset = load_dataset(RealFqpr())
-    multibeam = dset.raw_ping[0].isel(time=0).expand_dims('time')
-    qf = multibeam.qualityfactor
-    tpu = dset.return_tpu_parameters(list(dset.xyzrph['waterline'].keys())[0])
-
-    surface_ss = multibeam.soundspeed
-    beampointingangle = multibeam.beampointingangle
-    corr_beam_angle = xr.DataArray(data=expected_corrected_beam_angles, dims=['time', 'beam'],
-                                   coords={'time': multibeam.time.values, 'beam': multibeam.beam.values})
-    x = xr.DataArray(data=expected_alongtrack, dims=['time', 'beam'], coords={'time': multibeam.time.values,
-                                                                              'beam': multibeam.beam.values})
-    z = xr.DataArray(data=expected_depth, dims=['time', 'beam'], coords={'time': multibeam.time.values,
-                                                                         'beam': multibeam.beam.values})
-
-    raw_attitude = dset.raw_att
-    roll = raw_attitude['roll'].interp_like(beampointingangle)
-
-    tvu, thu = calculate_tpu(roll, beampointingangle, corr_beam_angle, x, z, surface_ss, tpu_dict=tpu,
-                             quality_factor=qf, vert_ref='waterline')
-    assert np.array_equal(thu, expected_thu)
-    assert np.array_equal(tvu, expected_tvu)
-
-
-def test_geohash():
-    newhash_vector = compute_geohash(np.array([43.123456789, 43.123456789, 43.123456789]),
-                                     np.array([-73.123456789, -73.123456789, -73.123456789]), precision=7)
-    newhash = _new_geohash(43.123456789, -73.123456789, precision=7)
-    assert newhash == newhash_vector[0]
-    assert newhash == b'drsj243'
-    lat, lon = decode_geohash(newhash)
-    assert lat == pytest.approx(43.12339782714844, abs=0.00000001)
-    assert lon == pytest.approx(-73.12294006347656, abs=0.00000001)
-
-
-def test_geohash_polygon():
-    polygon_test = np.array([[-70.1810536, 42.0519741], [-70.178872, 42.0501041], [-70.1813097, 42.0471989],
-                             [-70.1835136, 42.0490578], [-70.1810536, 42.0519741]])
-    innerhash, intersecthash = polygon_to_geohashes(polygon_test, precision=7)
-    assert innerhash == [b'drqp4yz']  # this cell is completely within the polygon
-    assert sorted(intersecthash) == sorted([b'drqp4yv', b'drqp4zp', b'drqp4zn', b'drqp5nc', b'drqp4yz', b'drqp4yx', b'drqp5nb', b'drqp5p0',
-                                            b'drqp4yr', b'drqp5n8', b'drqp5p2', b'drqp4yw', b'drqp4zr', b'drqp4yy', b'drqp5p1'])
