+++ HSTB/tests/resources/test_raster.tif.aux.xml
