+++ HSTB/tests/resources/test_shapefile.prj
