+++ HSTB/tests/resources/test_shapefile.shx
